services/orders/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.16.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
slowapi==0.1.9
redis==5.1.1
aiokafka==0.11.0
celery==5.4.0
cryptography==43.0.3
